collaboratory_website/requirements.txt
asgiref==3.3.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.1.7
django-filter==2.4.0
django-braces==1.14.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-heroku==0.3.1
django-import-export==2.5.0
djangorestframework==3.12.2
drf-writable-nested==0.6.2
et-xmlfile==1.0.1
MarkupPy==1.14
mysqlclient==2.0.3
odfpy==1.4.1
openpyxl==3.0.7
pytz==2021.1
PyYAML==5.4.1
sqlparse==0.4.1
tablib==3.0.0
xlrd==2.0.1
xlwt==1.3.0
